Encapsulated abstract class
